added parallelism example
