+++ test/runtests.jl
@@ -80,4 +80,4 @@ end
 end
 
 include("MOI_conshdlr.jl")
-include("cutsel_examples.jl")
+include("cutsel.jl")
